codereview: move non-tasks method to libraries
