+++ tasks/libs/common/gitlab.py
@@ -2,8 +2,10 @@ import json
 import os
 import platform
 import subprocess
+from collections import UserList
 from urllib.parse import quote
 
+import yaml
 from invoke.exceptions import Exit
 
 from tasks.libs.common.remote_api import APIError, RemoteAPI
@@ -353,3 +355,55 @@ def get_gitlab_bot_token():
         )
         raise Exit(code=1)
     return os.environ["GITLAB_BOT_TOKEN"]
+
+
+class ReferenceTag(yaml.YAMLObject):
+    """
+    Custom yaml tag to handle references in gitlab-ci configuration
+    """
+
+    yaml_tag = u'!reference'
+
+    def __init__(self, references):
+        self.references = references
+
+    @classmethod
+    def from_yaml(cls, loader, node):
+        return UserList(loader.construct_sequence(node))
+
+    @classmethod
+    def to_yaml(cls, dumper, data):
+        return dumper.represent_sequence(cls.yaml_tag, data.data, flow_style=True)
+
+
+def generate_gitlab_full_configuration(input_file):
+    """
+    Generate a full gitlab-ci configuration by resolving all includes
+    """
+    # Update loader/dumper to handle !reference tag
+    yaml.SafeLoader.add_constructor(ReferenceTag.yaml_tag, ReferenceTag.from_yaml)
+    yaml.SafeDumper.add_representer(UserList, ReferenceTag.to_yaml)
+
+    yaml_contents = []
+    read_includes(input_file, yaml_contents)
+    full_configuration = {}
+    for yaml_file in yaml_contents:
+        full_configuration.update(yaml_file)
+    return yaml.safe_dump(full_configuration)
+
+
+def read_includes(yaml_file, includes):
+    """
+    Recursive method to read all includes from yaml files and store them in a list
+    """
+    with open(yaml_file) as f:
+        current_file = yaml.safe_load(f)
+    if 'include' not in current_file:
+        includes.append(current_file)
+    else:
+        for include in current_file['include']:
+            if include.startswith('http'):
+                continue
+            read_includes(include, includes)
+        del current_file['include']
+        includes.append(current_file)
