+++ tasks/linter.py
@@ -1,9 +1,8 @@
 import re
 import sys
-from collections import UserList, defaultdict
+from collections import defaultdict
 from typing import List
 
-import yaml
 from invoke import Exit, task
 
 from tasks.build_tags import compute_build_tags_for_flavor
@@ -352,55 +351,3 @@ def is_get_parameter_call(file):
                         return SSMParameterCall(file, nb, with_wrapper=True, with_env_var=False)
         except UnicodeDecodeError:
             pass
-
-
-class ReferenceTag(yaml.YAMLObject):
-    """
-    Custom yaml tag to handle references in gitlab-ci configuration
-    """
-
-    yaml_tag = u'!reference'
-
-    def __init__(self, references):
-        self.references = references
-
-    @classmethod
-    def from_yaml(cls, loader, node):
-        return UserList(loader.construct_sequence(node))
-
-    @classmethod
-    def to_yaml(cls, dumper, data):
-        return dumper.represent_sequence(cls.yaml_tag, data.data, flow_style=True)
-
-
-def generate_gitlab_full_configuration(input_file):
-    """
-    Generate a full gitlab-ci configuration by resolving all includes
-    """
-    # Update loader/dumper to handle !reference tag
-    yaml.SafeLoader.add_constructor(ReferenceTag.yaml_tag, ReferenceTag.from_yaml)
-    yaml.SafeDumper.add_representer(UserList, ReferenceTag.to_yaml)
-
-    yaml_contents = []
-    read_includes(input_file, yaml_contents)
-    full_configuration = {}
-    for yaml_file in yaml_contents:
-        full_configuration.update(yaml_file)
-    return yaml.safe_dump(full_configuration)
-
-
-def read_includes(yaml_file, includes):
-    """
-    Recursive method to read all includes from yaml files and store them in a list
-    """
-    with open(yaml_file) as f:
-        current_file = yaml.safe_load(f)
-    if 'include' not in current_file:
-        includes.append(current_file)
-    else:
-        for include in current_file['include']:
-            if include.startswith('http'):
-                continue
-            read_includes(include, includes)
-        del current_file['include']
-        includes.append(current_file)
