+++ tasks/unit-tests/gitlab_api_tests.py
@@ -4,7 +4,7 @@ from unittest import mock
 
 from invoke.exceptions import Exit
 
-from tasks.libs.common.gitlab import Gitlab, get_gitlab_token
+from tasks.libs.common.gitlab import Gitlab, generate_gitlab_full_configuration, get_gitlab_token, read_includes
 from tasks.libs.common.remote_api import APIError
 
 
@@ -92,3 +92,31 @@ class TestStatusCode5XX(unittest.TestCase):
             failed = True
         if not failed:
             Exit("GitlabAPI was expected to fail")
+
+
+class TestReadIncludes(unittest.TestCase):
+    def test_with_includes(self):
+        includes = []
+        read_includes("tasks/unit-tests/testdata/in.yml", includes)
+        self.assertEqual(len(includes), 4)
+
+    def test_without_includes(self):
+        includes = []
+        read_includes("tasks/unit-tests/testdata/b.yml", includes)
+        self.assertEqual(len(includes), 1)
+
+
+class TestGenerateGitlabFullConfiguration(unittest.TestCase):
+    def test_nominal(self):
+        full_configuration = generate_gitlab_full_configuration("tasks/unit-tests/testdata/in.yml")
+        with open("tasks/unit-tests/testdata/out.yml") as f:
+            expected = f.read()
+        self.assertEqual(full_configuration, expected)
+
+    def test_yaml_with_reference(self):
+        full_configuration = generate_gitlab_full_configuration(
+            "tasks/unit-tests/testdata/ci_config_with_reference.yml"
+        )
+        with open("tasks/unit-tests/testdata/expected.yml") as f:
+            expected = f.read()
+        self.assertEqual(full_configuration, expected)
